代码更新，bug处理

+++ src/main/java/com/springboot/modules/sys/controller/AbstractController.java
@@ -19,7 +19,7 @@ public abstract class AbstractController {
         return (UserEntity)SecurityUtils.getSubject().getPrincipals();
     }
 
-    protected Integer getUserId(){
+    protected Long getUserId(){
         return getUser().getId();
     }
 }
